# api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "UWS: Unicorn Web Services"
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /instances:
    x-swagger-router-controller: instances
    get:
      tags:
        - Instances
      description: Returns list of instances
      operationId: list_instances
      parameters:
        - name: id
          type: string
          in: query
        - name: name
          type: string
          in: query
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Instance"
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
    post:
      tags:
        - Instances
      description: Creates a instance
      operationId: create_instance
      parameters:
        - name: body
          in: body
          schema:
            properties:
              name:
                type: string 
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/Instance"
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /instances/{id}:
    x-swagger-router-controller: instances
    get:
      tags:
        - Instances
      description: Gets information on a single instance
      operationId: read_instance
      parameters:
        - name: id
          in: path
          description: The ID of the instance
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Instance"
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
    put:
      tags:
        - Instances
      description: Updates a single instance
      operationId: update_instance
      parameters:
        - name: id
          in: path
          description: The ID of the instance
          required: true
          type: string
        - name: body
          in: body
          schema:
            properties:
              name:
                type: string 
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Instance"
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
    delete:
      tags:
        - Instances
      description: Deletes an instance
      operationId: delete_instance
      parameters:
        - name: id
          in: path
          description: The ID of the instance
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/GeneralResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  Instance:
    required:
      - id
      - name
    properties:
      id:
        type: string
      name:
        type: string
  GeneralResponse:
    required:
      - success
      - message
    properties:
      success:
        type: boolean
      message:
        type: string